+++ example/lib/main.dart
@@ -16,7 +16,7 @@ void main() async {
   // Server url for local backend on Android emulator / iOS simulator:
   // serverUrl: defaultTargetPlatform == TargetPlatform.android ? 'http://10.0.2.2:3000' : 'http://127.0.0.1:3000'
 
-  await PocketSync.instance.initialize(
+  await PocketSync.initialize(
     dbPath: path,
     options: PocketSyncOptions(
       projectId: projectId,